Backend/requirements.txt
IPykernel
python-pptx
langchain
langchain_community
# langchain_openai
langchain_google_genai
fastapi
uvicorn
python-multipart
